AC-3006: Split SplunkJS service into type and builder
