+++ ui/projects/setup/src/splunk/splunk-js.d.ts
@@ -1,6 +1,3 @@
-import { isDevMode } from "@angular/core";
-import { environment } from "../environments/environment";
-
 export type AllOrString = "-" | string;
 
 export type NamespaceSharing = "user" | "app" | "global" | "system";
@@ -102,56 +99,4 @@ export interface SplunkJs {
   Service: SplunkJsService;
 
   noConflict(): SplunkJs;
-}
-
-class SplunkJsProvider {
-  private static splunkJs: SplunkJs | undefined = undefined;
-
-  static get(): SplunkJs {
-    if (this.splunkJs) {
-      return this.splunkJs;
-    }
-
-    if (isDevMode()) {
-      // @ts-ignore
-      this.splunkJs = globalThis.splunkjs;
-    } else {
-      // @ts-ignore
-      this.splunkJs = globalThis.splunkjs.noConflict();
-    }
-
-    return this.splunkJs!;
-  }
-}
-
-export class SplunkJsServiceBuilder {
-  private service: SplunkJsService | undefined = undefined;
-
-  constructor(readonly namespace: Namespace) {}
-
-  getService(): SplunkJsService {
-    if (this.service) {
-      return this.service;
-    }
-
-    const splunkJs = SplunkJsProvider.get();
-
-    if (isDevMode()) {
-      const http = new splunkJs.ProxyHttp(environment.splunkManagementUrl);
-
-      const parameters = {
-        ...this.namespace,
-        username: environment.splunkManagementUsername,
-        password: environment.splunkManagementPassword,
-      };
-
-      this.service = new splunkJs.Service(http, parameters);
-    } else {
-      const http = new splunkJs.SplunkWebHttp();
-
-      this.service = new splunkJs.Service(http, this.namespace);
-    }
-
-    return this.service!;
-  }
 }
